chore: update the uniswap subgraph from v2 to v3

+++ lib/data/erc20/contract_service.dart
@@ -7,7 +7,7 @@ typedef TransferEvent = void Function(
     BigInt value);
 
 typedef TransferValue = void Function(
-    double? value);
+    BigInt? value);
 
 enum EventABIType {
   Transfer,
